Refactor current code, fix issue with game sometimes not starting
