+++ .gitignore
@@ -11,3 +11,4 @@ buildNumber.properties
 .mvn/wrapper/maven-wrapper.jar
 .vscode/
 generated_js/
+dependency-reduced-pom.xml
